Final Programacion III

+++ js/CargarPaginaPrincipal.js
@@ -1,11 +1,11 @@
 window.onload = function(){
-    let id = localStorage.getItem("idAdministrador");
+    let id = localStorage.getItem("idEmpleado");
     let funcionAjax = $.ajax({
     method : 'GET',
-    url : '../vendor/Login/TraerAdministrador/'+id
+    url : '../vendor/Login/TraerEmpleado/'+id
 });
     funcionAjax.then(function(dato){
-     document.getElementById("usuario").innerHTML = "<p id='usuario'><span class='glyphicon glyphicon-user'>"+dato.mail+"</span></p>";
+     document.getElementById("usuario").innerHTML = "<p id='usuario'><span class='glyphicon glyphicon-user'>"+dato.nombre+" "+dato.apellido+"</span></p>";
     },function(dato){
         console.log("No se pudo cargar el administrador!");
     });
@@ -25,10 +25,29 @@ function CerrarSesion()
   cancelButtonClass: 'btn btn-danger',
   cancelButtonText: 'No, no deseo cerrar sesión!'
 }).then(function () {
-  swal('Usted ha cerrado su sesión!').then(function(){
-      localStorage.clear();
-      window.location.replace("../enlaces/login.html");
+      //FALTA AJAX CON SESION - REGISTRAR SU SALIDA
+      let id = localStorage.getItem("idEmpleado");
+      var funcionAjax = $.ajax({
+      method : 'POST',
+      url : '../vendor/Login/CerrarSesion',
+      data : {idEmpleado : id}
+    });
+      funcionAjax.then(function(dato){
+         if(dato.status == 200)
+         {
+              swal('Usted ha cerrado su sesión!').then(function(){
+                  localStorage.clear();
+                  window.location.replace("../enlaces/login.html");
+              },function(){
+                 swal("OCURRIO ALGO INESPERADO!");
+              });
+         }
+         else if(dato.status == 400)
+         {
+          swal("Hubo un error al cerrar sesión del usuario!");
+         }
+      },function(dato){
+       console.log("ERROR en la API "+dato);
+      });
   });
-});
-
 }
